# cloud/openapi.yaml
openapi: 3.1.0
info:
  title: Jarvis Tax Assistant API
  description: |
    AI-powered tax and compliance assistant API for Indian tax professionals.
    Covers GST, Income Tax, Customs, Company Law, and FEMA.
    
    **Authentication**: Bearer JWT token or X-API-Key header.
    **License**: AGPL-3.0 (open source) / Commercial (enterprise features).
  version: 1.0.0
  contact:
    email: api@jarvis-tax.ai
  license:
    name: AGPL-3.0
    url: https://www.gnu.org/licenses/agpl-3.0.html
servers:
  - url: http://localhost:3001
    description: Development server
...
paths:
...
  /api/v1/tax/chat:
    post:
      tags: [Tax]
      summary: Ask a tax question
      security: [{ bearerAuth: [] }, { apiKeyAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [query]
              properties:
                query: { type: string, maxLength: 2000, example: "What is the GST rate for IT services?" }
                domain: { type: string, enum: [gst, income-tax, customs, company-law, fema, general], default: general }
                context: { type: array, items: { type: object, properties: { role: { type: string }, content: { type: string } } } }
                sessionId: { type: string }
      responses:
        '200':
          description: Tax answer with citations
          content:
            application/json:
              schema:
                type: object
                properties:
                  answer: { type: string }
                  references: { type: array, items: { type: object, properties: { type: { type: string }, title: { type: string }, citation: { type: string } } } }
                  confidence: { type: number }
                  domain: { type: string }
        '429':
          description: Daily query limit reached
  /api/v1/tax/compliance-calendar:
    get:
      tags: [Tax]
      summary: Get compliance deadlines
      security: [{ bearerAuth: [] }, { apiKeyAuth: [] }]
      parameters:
        - name: month
          in: query
          schema: { type: integer, minimum: 1, maximum: 12 }
        - name: year
          in: query
          schema: { type: integer }
        - name: category
          in: query
          schema: { type: string, enum: [gst, income-tax, tds, company-law, fema] }
      responses:
        '200':
          description: List of compliance deadlines
  /api/v1/tax/tariff-lookup:
    get:
      tags: [Tax]
      summary: Look up customs tariff rates
      security: [{ bearerAuth: [] }, { apiKeyAuth: [] }]
      parameters:
        - name: hsn
          in: query
          schema: { type: string }
          description: HSN code prefix
        - name: search
          in: query
          schema: { type: string }
          description: Description keyword search
        - name: chapter
          in: query
          schema: { type: string }
          description: Chapter name/number
      responses:
        '200':
          description: Tariff lookup results
  /api/v1/tax/analyze-document:
    post:
      tags: [Tax]
      summary: Analyze a tax document
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [action, fileType]
              properties:
                action: { type: string, enum: [analyze, compare, extract, validate] }
                fileType: { type: string, enum: [excel, pdf, word, csv] }
                analysisType: { type: string }
      responses:
        '200':
          description: Analysis results
        '403':
          description: Feature not available on current plan
  /api/v1/upload/tariff-data:
    post:
      tags: [Upload]
      summary: Upload customs tariff Excel data
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file: { type: string, format: binary }
                sheet: { type: string }
                year: { type: integer }
      responses:
        '200':
          description: Tariff data loaded successfully
  /api/v1/upload/tariff-search:
    get:
      tags: [Upload]
      summary: Search loaded tariff data
      security: [{ bearerAuth: [] }]
      parameters:
        - name: hsn
          in: query
          schema: { type: string }
        - name: search
          in: query
          schema: { type: string }
        - name: chapter
          in: query
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer, default: 50 }
      responses:
        '200':
          description: Search results
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key